+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Correctly display counters on console status page.
 - Correctly display friend names on console user details page.
 - Correctly display group names on console user details page.
+- Do not attempt to navigate to groups from console user details page.
 
 ## [2.5.0] - 2019-04-25
 ### Added